Migrate styles to registration
