+++ src/components/Form/_Form.scss
@@ -1,9 +1,9 @@
 .register-form {
-  height: calc(100% - 118px);
-  width: 100%;
-  position: relative;
-  display: flex;
-  flex-direction: column;
+  height: 100%;
+  // width: 100%;
+  // position: relative;
+  // display: flex;
+  // flex-direction: column;
 
   &__header {
     margin-bottom: 30px;
